+++ rs/pocket_ic_server/tests/test.rs
@@ -2,14 +2,22 @@ mod common;
 
 use crate::common::raw_canister_id_range_into;
 use candid::{Encode, Principal};
-use pocket_ic::common::rest::SubnetConfigSet;
+use ic_agent::agent::http_transport::ReqwestTransport;
+use ic_utils::interfaces::ManagementCanister;
+use pocket_ic::common::rest::{HttpsConfig, SubnetConfigSet};
+use pocket_ic::nonblocking::PocketIc as NonblockingPocketIc;
 use pocket_ic::{PocketIc, PocketIcBuilder};
+use rcgen::{CertificateParams, KeyPair};
 use reqwest::blocking::Client;
+use reqwest::Client as NonblockingClient;
 use reqwest::{StatusCode, Url};
 use std::io::Read;
+use std::io::Write;
+use std::net::{IpAddr, Ipv6Addr, SocketAddr};
 use std::path::PathBuf;
 use std::process::{Child, Command};
 use std::time::{Duration, Instant};
+use tempfile::NamedTempFile;
 
 pub const LOCALHOST: &str = "127.0.0.1";
 
@@ -198,74 +206,163 @@ fn test_port_file() {
     }
 }
 
-#[test]
-fn test_http_gateway() {
-    use ic_utils::interfaces::ManagementCanister;
-
-    // create live PocketIc instance
-    let mut pic = PocketIcBuilder::new()
-        .with_nns_subnet()
-        .with_application_subnet()
-        .build();
-    let endpoint = pic.make_live(None);
-    let app_subnet = pic.topology().get_app_subnets()[0];
-    let effective_canister_id =
-        raw_canister_id_range_into(&pic.topology().0.get(&app_subnet).unwrap().canister_ranges[0])
-            .start;
-
-    // deploy II canister to PocketIc instance proxying through HTTP gateway
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
+async fn test_gateway(
+    mut pic: NonblockingPocketIc,
+    client: NonblockingClient,
+    proto: &str,
+    port: u16,
+    expected_canister_id: Principal,
+) {
+    // create agent with custom transport
+    let transport = ReqwestTransport::create_with_client(
+        format!("{}://localhost:{}", proto, port),
+        client.clone(),
+    )
+    .unwrap();
+    let agent = ic_agent::Agent::builder()
+        .with_transport(transport)
         .build()
         .unwrap();
-    let ii_canister_id = rt.block_on(async {
-        let agent = ic_agent::Agent::builder()
-            .with_url(endpoint.clone())
-            .build()
-            .unwrap();
-        agent.fetch_root_key().await.unwrap();
+    agent.fetch_root_key().await.unwrap();
 
-        let ic00 = ManagementCanister::create(&agent);
-
-        let (canister_id,) = ic00
-            .create_canister()
-            .as_provisional_create_with_amount(None)
-            .with_effective_canister_id(effective_canister_id)
-            .call_and_wait()
-            .await
-            .unwrap();
-
-        let ii_path =
-            std::env::var_os("II_WASM").expect("Missing II_WASM (path to II wasm) in env.");
-        let ii_wasm = std::fs::read(ii_path).expect("Could not read II wasm file.");
-        ic00.install_code(&canister_id, &ii_wasm)
-            .with_raw_arg(Encode!(&()).unwrap())
-            .call_and_wait()
-            .await
-            .unwrap();
-
-        canister_id
-    });
-
-    // perform frontend asset request for the title page
-    let ii_url = endpoint
-        .join(&format!("?canisterId={}", ii_canister_id))
+    // deploy II canister to PocketIC instance using agent and proxying through HTTP(S) gateway
+    let ic00 = ManagementCanister::create(&agent);
+    let (canister_id,) = ic00
+        .create_canister()
+        .as_provisional_create_with_amount(None)
+        .with_effective_canister_id(expected_canister_id)
+        .call_and_wait()
+        .await
         .unwrap();
-    let client = Client::new();
-    let res = client.get(ii_url.clone()).send().unwrap();
-    let page = String::from_utf8(res.bytes().unwrap().to_vec()).unwrap();
+    assert_eq!(canister_id, expected_canister_id);
+
+    // install II canister WASM
+    let ii_path = std::env::var_os("II_WASM").expect("Missing II_WASM (path to II wasm) in env.");
+    let ii_wasm = std::fs::read(ii_path).expect("Could not read II wasm file.");
+    ic00.install_code(&canister_id, &ii_wasm)
+        .with_raw_arg(Encode!(&()).unwrap())
+        .call_and_wait()
+        .await
+        .unwrap();
+
+    // perform frontend asset request for the title page at http://localhost:<port>/?canisterId=<canister-id>
+    let canister_url = format!("{}://localhost:{}/?canisterId={}", proto, port, canister_id);
+    let res = client.get(canister_url).send().await.unwrap();
+    let page = String::from_utf8(res.bytes().await.unwrap().to_vec()).unwrap();
+    assert!(page.contains("<title>Internet Identity</title>"));
+
+    // perform frontend asset request for the title page at http://<canister-id>.localhost:<port>
+    let canister_url = format!("{}://{}.localhost:{}", proto, canister_id, port);
+    let res = client.get(canister_url.clone()).send().await.unwrap();
+    let page = String::from_utf8(res.bytes().await.unwrap().to_vec()).unwrap();
     assert!(page.contains("<title>Internet Identity</title>"));
 
     // stop HTTP gateway and disable auto progress
-    pic.stop_live();
+    pic.stop_live().await;
 
     // HTTP gateway requests should eventually stop and requests to it fail
     loop {
-        if client.get(ii_url.clone()).send().is_err() {
+        if client.get(canister_url.clone()).send().await.is_err() {
             break;
         }
         std::thread::sleep(Duration::from_millis(20));
     }
+}
+
+#[tokio::test]
+async fn test_http_gateway() {
+    // create PocketIc instance
+    let mut pic = PocketIcBuilder::new()
+        .with_nns_subnet()
+        .with_application_subnet()
+        .build_async()
+        .await;
+
+    // retrieve the first canister ID on the application subnet
+    // which will be the effective and expected canister ID for canister creation
+    let topology = pic.topology().await;
+    let app_subnet = topology.get_app_subnets()[0];
+    let effective_canister_id =
+        raw_canister_id_range_into(&topology.0.get(&app_subnet).unwrap().canister_ranges[0]).start;
+
+    // make PocketIc instance live with an HTTP gateway
+    let port = pic.make_live(None).await.port_or_known_default().unwrap();
+
+    // create a non-blocking reqwest client with resolving localhost and <canister-id>.localhost to [::1]
+    let client = NonblockingClient::builder()
+        .resolve(
+            "localhost",
+            SocketAddr::new(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)), port),
+        )
+        .resolve(
+            &format!("{}.localhost", effective_canister_id),
+            SocketAddr::new(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)), port),
+        )
+        .build()
+        .unwrap();
+
+    test_gateway(pic, client, "http", port, effective_canister_id.into()).await;
+}
+
+#[tokio::test]
+async fn test_https_gateway() {
+    // create PocketIc instance
+    let mut pic = PocketIcBuilder::new()
+        .with_nns_subnet()
+        .with_application_subnet()
+        .build_async()
+        .await;
+
+    // retrieve the first canister ID on the application subnet
+    // which will be the effective and expected canister ID for canister creation
+    let topology = pic.topology().await;
+    let app_subnet = topology.get_app_subnets()[0];
+    let effective_canister_id =
+        raw_canister_id_range_into(&topology.0.get(&app_subnet).unwrap().canister_ranges[0]).start;
+
+    // generate root TLS certificate
+    let root_key_pair = KeyPair::generate().unwrap();
+    let localhost = "localhost";
+    let canister_localhost = &format!("{}.localhost", effective_canister_id);
+    let root_cert =
+        CertificateParams::new(vec![localhost.to_string(), canister_localhost.to_string()])
+            .unwrap()
+            .self_signed(&root_key_pair)
+            .unwrap();
+    let (mut cert_file, cert_path) = NamedTempFile::new().unwrap().keep().unwrap();
+    cert_file.write_all(root_cert.pem().as_bytes()).unwrap();
+    let (mut key_file, key_path) = NamedTempFile::new().unwrap().keep().unwrap();
+    key_file
+        .write_all(root_key_pair.serialize_pem().as_bytes())
+        .unwrap();
+
+    // make PocketIc instance live with an HTTPS gateway
+    let https_config = HttpsConfig {
+        cert_path: cert_path.into_os_string().into_string().unwrap(),
+        key_path: key_path.into_os_string().into_string().unwrap(),
+    };
+    let port = pic
+        .make_live_https(None, localhost.to_string(), https_config)
+        .await
+        .port_or_known_default()
+        .unwrap();
+
+    // create a non-blocking reqwest client with a custom root certificate
+    // and resolving localhost and <canister-id>.localhost to [::1]
+    let client = NonblockingClient::builder()
+        .add_root_certificate(reqwest::Certificate::from_pem(root_cert.pem().as_bytes()).unwrap())
+        .resolve(
+            localhost,
+            SocketAddr::new(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)), port),
+        )
+        .resolve(
+            canister_localhost,
+            SocketAddr::new(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)), port),
+        )
+        .build()
+        .unwrap();
+
+    test_gateway(pic, client, "https", port, effective_canister_id.into()).await;
 }
 
 #[test]