+++ packages/pocket-ic/CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The library function `PocketIc::create_canister_with_id` works for all IC mainnet canister IDs that do not belong to the NNS or II subnet.
 - The library function `PocketIc::uninstall_canister` to uninstall code of an existing canister.
 - The library function `PocketIc::update_canister_settings` to update settings (e.g., compute allocation) of an existing canister.
+- The library function `PocketIc::make_live_https` configuring a PocketIc instance to automatically make progress (updating time and executing rounds)
+  and creating an HTTPS gateway for that instance listening at a dedicated domain and port and using a specified TLS certificate.
 
 ### Removed
 - Public field `instance_id` in the synchronous PocketIc library, use the function `instance_id` instead